Split Incubation APIs from BackpackController
