+++ backend/src/routes/Backpack.js
@@ -2,10 +2,6 @@ import { Router } from "express"
 import {
 	getItemsFromBackpack,
 	useItemsOnMonster,
-	getIncubatingEggs,
-	getIncubatingEggByUId,
-	incubateAnEgg,
-	hatchAnEgg,
 } from "../controllers/BackpackController.js"
 import validateJwt from "../middleware/JwtTokenHandler.js"
 
@@ -14,9 +10,5 @@ const router = Router()
 router
 	.get("/backpack", validateJwt, getItemsFromBackpack)
 	.put("/backpack/use", validateJwt, useItemsOnMonster)
-	.get("/backpack/incubation", validateJwt, getIncubatingEggs)
-	.post("/backpack/incubate", validateJwt, incubateAnEgg)
-	.post("/backpack/hatch", validateJwt, hatchAnEgg)
-	.get("/backpack/incubation/:incubation_uid", validateJwt, getIncubatingEggByUId)
 
 export default router
