+++ backend/src/routes/api.js
@@ -4,6 +4,7 @@ import BackpackRouter from "./Backpack.js"
 import AuthenticationRouter from "./Authentication.js"
 import TrainerRouter from "./Trainer.js"
 import ChallengeRouter from "./Challenge.js"
+import IncubationRouter from "./Incubation.js"
 
 const apiRouter = Router()
 apiRouter.get("/", (req, res) => {
@@ -15,5 +16,6 @@ apiRouter.use(TrainerRouter)
 apiRouter.use(MonsterRouter)
 apiRouter.use(BackpackRouter)
 apiRouter.use(ChallengeRouter)
+apiRouter.use(IncubationRouter)
 
 export default apiRouter
